Atualização das rotas get e delete

+++ routes/routes.js
@@ -3,27 +3,30 @@ var router  = express.Router();
 const todoService = require('../helpers/todoService');
 const userService = require('../helpers/userService');
 
+const { validaRegistro } = require("../middlewares/registro");
+const { updateUser } = require("../middlewares/update");
+
 //Rota inicial
 router.get('/', (req, res) => {
     res.status(200).json({msg: "Working api routes!"})
-})
+});
 
 //Rota para registro de user
-router.post('/registrar', async (req, res) => {
+router.post('/registrar', validaRegistro, async (req, res) => {
     const nome = req.body.name
     console.log(nome)
     try {
         const newUser = await userService.new({name: nome});
         console.log(newUser)
-        return res.status(200).json({ msg: "Cadastrado", user: newUser });
+        return res.status(201).json({ msg: "Cadastrado", user: newUser });
     } catch (error) {
         console.log(error);
         return res.status(500).json({ msg: "Um erro aconteceu!" });
     }
-})
+});
 
 //Rota para modificação de user
-router.put('/update/:nomeUserAntigo/:nomeUserUpdated', async(req, res) => {
+router.put('/update/:nomeUserAntigo/:nomeUserUpdated', updateUser, async(req, res) => {
     const {nomeUserAntigo, nomeUserUpdated} = req.params;
 
     try {
@@ -36,32 +39,41 @@ router.put('/update/:nomeUserAntigo/:nomeUserUpdated', async(req, res) => {
         console.log(error);
         return res.status(500).json({ msg: "Um erro aconteceu!" });
     }
-
-})
+});
 
 //Rota para obter um user
 router.get('/listUser/:id', async (req, res) => {
+   try {
     const foundUser = await userService.findByPK(req.params.id);
-    if(foundUser) {
-        return res.status(200).json({msg: "User encontrado!", user: foundUser});
-    } else{
-        return res.status(404).json({ msg: "User não encontrado" });
+        if(foundUser) {
+            return res.status(200).json({msg: "User encontrado!", user: foundUser});
+        } else{
+            return res.status(404).json({ msg: "User não encontrado" });
+        }
+    } catch (error) {
+        return res.status(500).json({ msg: "Um erro aconteceu." });
     }
-})
+});
 
 //Rota para exclusão de user
 router.delete('/delete/:id', async(req, res) => {
     const id = req.params.id
     
     try {
-        const deleted = await userService.delete(id);
-        return res.status(200).json({ msg: "Deletado!"});
+        const user = await userService.findByPK(id);
+
+        if (user) {
+            await userService.delete(id);
+            return res.status(200).json({ msg: "Deletado!"});
+        } else {
+            return res.status(400).json({ msg: "Usuário não encotrado!"});
+        }
+        
     } catch (error) {
         console.log(error);
         return res.status(500).json({ msg: "Um erro aconteceu!" });
     }
-
-})
+});
 
 //Rota para adição de uma tarefa
 router.post("/newToDo", async (req, res) => {
@@ -84,8 +96,7 @@ router.post("/newToDo", async (req, res) => {
         console.log(error);
         return res.status(404).json({msg: "Um erro aconteceu!", error: error})
     }
-    
-})
+});
 
 //Rota para exclusão de uma tarefa
 router.delete('/deleteToDo/:userId/:id', async(req, res) => {
@@ -98,8 +109,7 @@ router.delete('/deleteToDo/:userId/:id', async(req, res) => {
         console.log(error);
         return res.status(404).json({msg: "Um erro aconteceu!", error: error})
     }
-    
-})
+});
 
 //Rota para listagem da tarefa de determinado usuário
 router.get('/allToDo/:userId', async(req, res) => {
@@ -112,14 +122,13 @@ router.get('/allToDo/:userId', async(req, res) => {
         console.log(error);
         return res.status(404).json({msg: "Um erro aconteceu!", error: error})
     }
-})
+});
 
 //Rota para atualização da tarefa
 router.put('/changeToDo/:userId/:id', async(req, res) => {
     const {userId, id} = req.params;
 
     const {titulo, descricao, dataEntrega} = req.body
-
 
     try{
         let change = todoService.updateById(userId, id, {
@@ -135,6 +144,6 @@ router.put('/changeToDo/:userId/:id', async(req, res) => {
         console.log(titulo)
         return res.status(404).json({msg: "Um erro aconteceu!", error: error})
     }
-})
+});
 
 module.exports = router;
